requirements.txt
fastapi
aiogram==2.22.1
gunicorn
uvicorn
python-dotenv
requests
stripe
